apparently i called Gear.load on weapon THREE TIMES per attack, that was insane

+++ src/main/java/com/robertx22/loot/LootUtils.java
@@ -70,7 +70,7 @@ public class LootUtils {
         float after_world = after_mob_health;
 
         if (world.isMapWorld()) {
-            after_world = after_mob_health * 1 + world.getMap()
+            after_world = after_mob_health * 1 + (float) world.getMap()
                     .getBonusLootAmount() / 100;
         }
         if (entity instanceof EntitySlime) {
